refactor: date range picker component
